Code review fixes
Remove whitelistedips package and put whitelist checking code in
IprepdIO. Revert optional namespace code.
